bbs comments v1

+++ bbs/src/main/webapp/WEB-INF/views/bbs/read.jsp
@@ -107,6 +107,7 @@
 
 	function getReplies() {
 		var replies="";
+		var userId = "${userId!=null?userId:'guest'}";
 		$.getJSON("/bbs/reply/${board.bno}", function(data){			
 			$(data).each(function() {
 				//alert(this.replyer);
@@ -118,16 +119,22 @@
 				replies += "<span class='text-muted'>"+
 							(new Date(this.moddate).toLocaleString('ko-KR', { timeZone: 'UTC' }))
 									+"</span></div>";
-				replies += "<div class='panel-body'>"+this.content+"</div></div></div>";
-				replies += "<div><button class='btn btn-warning'>수정</button> ";
-				replies += "<button class='btn btn-danger' onclick='deleteReply("+this.rno+")'>";
-				replies += "삭제</button></div></div>";			
+				replies += "<div class='panel-body' id='rc"+this.rno+"'>"+this.content+"</div></div></div>";
+				if(this.replyer == userId) {
+					replies += "<div><button class='btn btn-warning' ";
+					replies += "onclick='updateReply("+this.rno+")'>수정</button> ";
+					replies += "<button class='btn btn-danger' onclick='deleteReply("+this.rno+")'>";
+					replies += "삭제</button></div>";
+				}
+				replies += "</div>";			
 			});
 			$("#replies").html(replies);
 		});
 	}
 
 	function deleteReply(rno) {
+		var ans = confirm("Are you sure to delete?");
+		if(ans==false) return;
 		$.ajax({
 			type:'delete',
 			url: '/bbs/reply/'+rno,
@@ -136,13 +143,19 @@
 				"X-HTTP-Method-Override": "DELETE"},
 			success: function(result) {
 				$("#"+rno).hide('slow');
+			}, 
+			error: function(result) {
+				alert("삭제할 수 없습니다[권한 확인].")
 			}
 		});
 	}
 
 	function addReply() {
 		var content = $("#addReply").val();
-		if (content=='') return;
+		if (content=='') {
+			alert('글을 입력하세요!');
+			return;
+		}
 		$.ajax({
 			type:'POST',
 			url: '/bbs/reply/',
@@ -154,11 +167,68 @@
 			success: function(result) {
 				getReplies();
 				 $("#addReply").val("");
+			},
+			error: function(result) {
+				alert("추가할 수 없습니다.");
 			}
 		});
-
 	}
-	
+
+	function updateReply(rno) {		
+		var orgCont = $('#rc'+rno).html();
+		$('#replyUpdateModal').modal('toggle');
+		$('#updateReplyContent').val(orgCont);
+		$('#replyIdToBeUpdated').val(rno);
+	}
+
+	function updateCommit() {
+		var replyer = "${userId!=null?userId:'guest'}";
+		var rno = $('#replyIdToBeUpdated').val();
+		var content = $('#updateReplyContent').val();
+
+		$.ajax({
+			type:'PUT', 
+			url: '/bbs/reply/'+rno, 
+			headers: {
+				'Content-Type': 'application/json', 
+				'X-HTTP-Method-Override': 'PUT'
+				}, 
+			data : JSON.stringify({'content':content, 'replyer':replyer}), 
+			dataType: 'text',
+			success:function(result) {
+				$('#rc'+rno).html(content);
+			},
+			error: function(result) {
+				alert("수정할 수 없습니다[권한 확인]!");
+			}
+		});
+		$('#replyUpdateModal').modal('hide');
+	}
 </script>
+
+<!-- Modal -->
+<div class="modal fade" id="replyUpdateModal" tabindex="-1" role="dialog" 
+					aria-labelledby="replyUpdateModalLabel" aria-hidden="true">
+  <div class="modal-dialog" role="document">
+    <div class="modal-content">
+      <div class="modal-header">
+        <h5 class="modal-title" id="replyUpdateModalLabel">댓글 수정</h5>
+        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+          <span aria-hidden="true">&times;</span>
+        </button>
+        <input type="hidden" name="rno" id="replyIdToBeUpdated">
+      </div>
+      <div class="modal-body">
+        <input type="text" class="form-control" id="updateReplyContent">
+      </div>
+      <div class="modal-footer">
+        <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
+        <button type="button" class="btn btn-primary" 
+         onclick="updateCommit()" id="updateCommit">Save changes</button>
+      </div>
+    </div>
+  </div>
+</div>
+
 </body>
 </html>
